Move provider creations to utils folder
